requirements/local.txt
absl-py==0.10.0
appnope==0.1.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
dm-tree==0.1.5
edward2 @ git+https://github.com/google/edward2.git@38ef7ce89e686271f0be0c9369cb73b9017f7fd4
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-resources==4.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.4
more-itertools==8.4.0
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.14.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
robustness-metrics @ git+https://github.com/google-research/robustness_metrics.git@b4c4d4d19e3ef0f2dde5aa44292f82eddeb86906
rsa==4.6
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-datasets==4.1.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.26.0
tensorflow-probability==0.12.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.55.1
traitlets==4.3.3
typing-extensions==3.7.4.3
uncertainty-metrics==0.0.81
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
scikit-image==0.18.1